+++ app/src/main/java/io/webthings/app/MainActivity.kt
@@ -1,4 +1,4 @@
-package io.webthings.webthings
+package io.webthings.app
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
@@ -10,7 +10,7 @@ import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import io.webthings.webthings.ui.theme.WebthingsTheme
+import io.webthings.app.ui.theme.WebthingsTheme
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
